refactor: mdx에서 html로 변환하면서 제목 태그 정보를 추출하는 코드를 post service에서 PostArticleContent 컴포넌트로 이동
