+++ src/components/posts/article/post-article-content.tsx
@@ -1,3 +1,6 @@
+import remarkHeadings, {
+  type Heading as RemarkHeading,
+} from "@vcarl/remark-headings";
 import { bundleMDX } from "mdx-bundler";
 import { getMDXComponent } from "mdx-bundler/client";
 import Link from "next/link";
@@ -10,10 +13,17 @@ import remarkGfm from "remark-gfm";
 import remarkHeadingId, {
   type RemarkHeadingIdOptions,
 } from "remark-heading-id";
+import remarkParse from "remark-parse";
 import remarkSectionize from "remark-sectionize";
+import remarkStringify from "remark-stringify";
+import { unified } from "unified";
 import { Heading } from "@/components/ui/heading";
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area";
-import { type Post } from "@/schema/posts";
+import {
+  PostContentHeading,
+  PostContentHeadingType,
+  type Post,
+} from "@/schema/posts";
 import { PostArticleTOC } from "./toc/post-article-toc";
 
 type Props = {
@@ -34,6 +44,46 @@ const remarkHeadingIdOptions: RemarkHeadingIdOptions = {
   uniqueDefaults: true, // 값이 같더라도 다른 id 값을 생성하도록 설정
 };
 
+function convertDepthToHeadingType(depth: number): PostContentHeadingType {
+  switch (depth) {
+    case 2:
+      return "h2";
+    case 3:
+      return "h3";
+    default: {
+      throw new Error(
+        "mdx의 제목의 depth는 2, 3만 가능합니다. (h2, h3만 허용합니다.)",
+      );
+    }
+  }
+}
+
+async function extractHeadingsFromMDXString(sourceMDXString: string) {
+  const processor = unified()
+    .use(remarkParse) // markdown -> syntax tree(mdast)
+    .use(remarkStringify) // syntax tree(mdast) -> markdown
+    .use(remarkHeadingId, remarkHeadingIdOptions) // markdown의 헤더(#, ##, ### ...)에 id attribute 주입(옵션을 이용해 값을 이용해 자동 설정)
+    .use(remarkHeadings); // markdown의 헤더(#, ##, ###) 데이터만 추출
+
+  const headings: PostContentHeading[] = [];
+
+  const headingsByRemarkHeadings = (await processor.process(sourceMDXString))
+    .data.headings as RemarkHeading[];
+
+  for (const headingByRemarkHeading of headingsByRemarkHeadings) {
+    const id = headingByRemarkHeading.data?.id;
+    if (!id || typeof id !== "string") {
+      throw new Error("mdx의 제목에 id 속성이 필요합니다.");
+    }
+
+    const { depth, value } = headingByRemarkHeading;
+
+    headings.push({ id, text: value, type: convertDepthToHeadingType(depth) });
+  }
+
+  return headings;
+}
+
 export async function PostArticleContent({ post }: Props) {
   const { code } = await bundleMDX({
     source: post.mdxContent,
@@ -53,6 +103,8 @@ export async function PostArticleContent({ post }: Props) {
     },
   });
 
+  const headings = await extractHeadingsFromMDXString(post.mdxContent);
+
   const MDXComponent = getMDXComponent(code);
 
   return (
@@ -65,7 +117,7 @@ export async function PostArticleContent({ post }: Props) {
         그래서 TOC 컴포넌트에 sticky를 준다.
       */}
       <aside className="lg:order-1 lg:ml-5 lg:border-l lg:border-l-border">
-        <PostArticleTOC headings={post.headings} />
+        <PostArticleTOC headings={headings} />
       </aside>
       <div
         id="article-content"
